requirements/base.txt
django==1.5.1
bpython==0.12
django-braces==1.0.0
django-model-utils==1.3.1
logutils==0.3.3
South==0.7.6
django-allauth==0.14.2
django-bootstrap-toolkit==2.11.0
django-json-field==0.5.1
django-categories==1.2.1
#-e git+https://github.com/toastdriven/django-tastypie.git#egg=tastypie
fabric==1.6.1
git+https://github.com/jezdez/django-avatar.git#egg=avatar
PIL==1.1.7
django-ses==0.4.1
django-haystack==2.0.0
django-celery==3.0.17
git+https://github.com/dcramer/django-paypal.git#egg=django-paypal
paypal==1.2.0
pycrypto==2.6
twilio==3.5.4
django-twilio==0.4